Add code to recalculate line item total
This commit adds logic to recalculate line item price once the user
is associated to the order, The current flow of code used to create the
line item and order once the product was added to cart.

To calculate accurate pricing for the user, we have added a new method
which is called after the user is associated to the order and based on
the conditions for calculating price based on user group the line item price
is recalculated.